previous web page was added with the remaining components. The previous issue was rectified.
The previous issue of the web page having the maximize and minimize window issue was rectified by adding a alert portal before the actual page opens. The alert contained a text saying the user to maximize the tab and click yes to proceed further to open the main page.

The remaining components of the main page were added and updated.